feat: yml changed and generated splode files added
