+++ Samples/Integers/mocks.c
@@ -1,0 +1,8 @@
+#include <stdio.h>
+#include <stdarg.h>
+#include "mocks.h"
+
+int print(const char *format, ...)
+{
+  return 0;
+}
