+++ Samples/Integers/mocks.h
@@ -1,0 +1,6 @@
+#ifndef MOCK_H
+#define MOCK_H
+
+void print(const char *format, ...);
+
+#endif
