+++ Samples/Integers/overflow_if/overflow_if_pedro_splode.c
@@ -1,6 +1,11 @@
 #include <limits.h>
 #include "../mocks.h"
 
+#include <klee/klee.h>
+
+// Ansatz file
+
+
 int pedro(int input)
 {
     int overflow_example = INT_MAX;
@@ -10,4 +15,18 @@ int pedro(int input)
         print("Overflow result: %d\n", overflow_example);
     }
     return 0;
+}
+
+// End ansatz file
+
+
+
+int main() {
+
+  int i;
+  klee_make_symbolic(&i, sizeof(i), "i");
+
+  pedro(i);
+
+  return 0;
 }
