+++ Samples/Memory/buffer_overflow_for/buffer_overflow_for_juan_splode.c
@@ -1,6 +1,9 @@
 #include <string.h>
 #include <stdlib.h>
 
+#include <klee/klee.h>
+
+// Ansatz file
 int CONSTANT;
 int juan(int size)
 {
@@ -14,4 +17,21 @@ int juan(int size)
     buffer[CONSTANT - 1] = '\0';
 
     return 0;
+}
+
+// End ansatz file
+
+
+
+int main() {
+
+  CONSTANT = 5;
+
+
+  int i;
+  klee_make_symbolic(&i, sizeof(i), "i");
+
+  juan(i);
+
+  return 0;
 }
