+++ Samples/Memory/double_free_if/double_free_if_jorge_splode.c
@@ -1,5 +1,8 @@
 #include <stdlib.h>
 
+#include <klee/klee.h>
+
+// Ansatz file
 int jorge(int condition)
 {
     int *ptr = malloc(sizeof(int));
@@ -11,4 +14,18 @@ int jorge(int condition)
     }
 
     return 0;
+}
+
+// End ansatz file
+
+
+
+int main() {
+
+  int i;
+  klee_make_symbolic(&i, sizeof(i), "i");
+
+  jorge(i);
+
+  return 0;
 }
