+++ Samples/Memory/mem_corruption_ternary/mem_corruption_switch.yml
@@ -1,14 +1,10 @@
 ---
-set-up:
-  includes:
-    - <stdio.h>
-    - <stdlib.h>
+prologue: |
+  #include <stdlib.h>
 
-mocks:
-  - printf:
-    return: int
-    args:
-      - const char*
-      - ... # printf uses 'ellipsis'
-    return-value: 0
-ansatz: jaime
+ansatz-call:
+  name: jaime
+  arguments:
+    - name: i
+      symbolic: true
+      type: int
