+++ Samples/Memory/mem_corruption_ternary/mem_corruption_switch_jaime_splode.c
@@ -1,5 +1,8 @@
 #include <stdlib.h>
 
+#include <klee/klee.h>
+
+// Ansatz file
 int jaime(int index)
 {
     int *array = malloc(5 * sizeof(int));
@@ -16,4 +19,18 @@ int jaime(int index)
 
     free(array);
     return 0;
+}
+
+// End ansatz file
+
+
+
+int main() {
+
+  int i;
+  klee_make_symbolic(&i, sizeof(i), "i");
+
+  jaime(i);
+
+  return 0;
 }
