+++ Samples/Memory/use_after_free_ternary/use_after_free_ternary.yml
@@ -1,19 +1,15 @@
 ---
-set-up:
-  includes:
-    - <stdio.h>
-    - <stdlib.h>
-  gloabls:
-    - name: CONSTANT
-      value: 10
-    - name: CONDITION
-      value: 1
+prologue: |
+  #include <string.h>
+  #include <stdlib.h>
 
-mocks:
-  - printf:
-    return: int
-    args:
-      - const char*
-      - ... # printf uses 'ellipsis'
-    return-value: 0
-ansatz: jose
+main-set-up: |
+  CONSTANT = 10;
+  CONDITION = 1;
+
+ansatz-call:
+  name: jose
+  arguments:
+    - name: i
+      symbolic: true
+      type: int
