+++ Samples/Memory/use_after_free_ternary/use_after_free_ternary_jose_splode.c
@@ -1,6 +1,11 @@
 #include <string.h>
 #include <stdlib.h>
 
+#include <klee/klee.h>
+
+// Ansatz file
+
+
 int CONSTANT;
 int CONDITION;
 int jose(int condition)
@@ -21,4 +26,22 @@ int jose(int condition)
     }
 
     return 0;
+}
+
+// End ansatz file
+
+
+
+int main() {
+
+  CONSTANT = 10;
+  CONDITION = 1;
+
+
+  int i;
+  klee_make_symbolic(&i, sizeof(i), "i");
+
+  jose(i);
+
+  return 0;
 }
